refactor: move response types into pkg response
