+++ helper.go
@@ -17,19 +17,6 @@ func verifyBase(route string) bool {
 	return true
 }
 
-func contentTypeOf(response interface{}) string {
-	switch response.(type) {
-	case Html:
-		return "text/html"
-	case Json:
-		return "application/json"
-	case string:
-		return "text/plain"
-	default:
-		return "text/plain"
-	}
-}
-
 func parseParameter(v reflect.Value, param string) reflect.Value {
 	switch v.Kind() {
 	case reflect.Bool:
