--- response_types.go
@@ -1,8 +1,0 @@
-package rocket
-
-// Html is a mark for text is HTML
-// return "Content-Type": "text/html"
-type Html string
-
-// Json will return "Content-Type": "application/json"
-type Json string
